Count cage stats from the cage memo instead of n^k brute force (#109)
Closes #108

## Problem

`Puzzle::cage_tuples` enumerated all `n^k` value combinations, checking
only the cells' current fills and collinear distinctness — never the
cage's arithmetic constraint:

- **Wrong counts**: an `Add(5)` domino in a 4×4 reported 12 tuples
instead of 4; the `+6` L-cage reported far more than its 7 valid tuples.
The existing tests passed only because they used fill-decomposable
cages.
- **Hangs on fat cages**: a 3×4 cage in a 9×9 is `9^12 ≈ 5×10^11`
iterations, and `CageStats` fired this on every selection change inside
a cage.

## Fix

Replace `cage_tuples` with `Puzzle::cage_viable_counts`, which narrows
the cage's memo — the exact tuple relation per ADR-0006 — by the current
grid fills and counts from the diagram:

- `Mdd::tuple_count`: memoized root→terminal path-count fold, `O(states
+ edges)`.
- `Mdd::multiset_count`: bottom-up fold of distinct suffix multisets per
node, layer by layer (every edge runs from depth `d` to `d+1`), bounded
by diagram size × multiset count, never the path count.
- Non-commutative cages count their explicit `Table` pair list; `Given`
is trivial.
- The designer's `PartialSolution::viable_counts` routes through the new
API; counts are `u64` so fat-cage path counts can't overflow 32-bit
`usize` on WASM.

## Tests

- Regression: `Add(5)` domino in 4×4 → `(2 multisets, 4 tuples)`, not
12; `+6` L-cage → `(2, 7)`, matching its `Mdd`.
- Counts agree with the `Mdd::tuples()` brute-force oracle across a grid
of `n`/`k`/operator/target combinations, cage shapes with collinear
lines, and after narrowing.
- Perf: an 8-cell `Add` cage in a 9×9 (previously `9^8` iterations; the
3×4 case never finished) now completes via the memo fold.
- `cargo clippy` clean; full workspace tests pass (the Tauri crate
doesn't build in this container for lack of GTK system libraries,
unrelated to this change).


---
_Generated by [Claude
Code](https://claude.ai/code/session_01RvWRACT5JaPzySyY9WH6g1)_

---------

+++ apps/designer/src/components/cage_stats.rs
@@ -33,7 +33,7 @@ pub fn CageStats() -> impl IntoView {
     }
 }
 
-fn pluralize(n: usize, singular: &str, plural: &str) -> String {
+fn pluralize(n: u64, singular: &str, plural: &str) -> String {
     if n == 1 {
         format!("1 {singular}")
     } else {
